knowledge graph/requirements.txt
spacy
pdfminer.six
pytesseract
py2neo
beautifulsoup4
requests
